+++ pkg/appdef/utils_filter.go
@@ -6,7 +6,7 @@
 package appdef
 
 import (
-	"testing"
+	"strings"
 
 	"github.com/voedger/voedger/pkg/coreutils/utils"
 )
@@ -35,62 +35,19 @@ func FirstFilterMatch(f IFilter, types SeqType) IType {
 	return nil
 }
 
-func TestFilterKind_MarshalText(t *testing.T) {
-	tests := []struct {
-		name string
-		k    FilterKind
-		want string
-	}{
-		{name: `0 —> "FilterKind_null"`,
-			k:    FilterKind_null,
-			want: `FilterKind_null`,
-		},
-		{name: `1 —> "FilterKind_QNames"`,
-			k:    FilterKind_QNames,
-			want: `FilterKind_QNames`,
-		},
-		{name: `FilterKind_count —> <number>`,
-			k:    FilterKind_count,
-			want: utils.UintToString(FilterKind_count),
-		},
+func (k FilterKind) MarshalText() ([]byte, error) {
+	var s string
+	if k < FilterKind_count {
+		s = k.String()
+	} else {
+		s = utils.UintToString(k)
 	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := tt.k.MarshalText()
-			if err != nil {
-				t.Errorf("FilterKind.MarshalText() unexpected error %v", err)
-				return
-			}
-			if string(got) != tt.want {
-				t.Errorf("FilterKind.MarshalText() = %s, want %v", got, tt.want)
-			}
-		})
-	}
-
-	t.Run("100% cover FilterKind.String()", func(t *testing.T) {
-		const tested = FilterKind_count + 1
-		want := "FilterKind(" + utils.UintToString(tested) + ")"
-		got := tested.String()
-		if got != want {
-			t.Errorf("(FilterKind_count + 1).String() = %v, want %v", got, want)
-		}
-	})
+	return []byte(s), nil
 }
 
-func TestFilterKindTrimString(t *testing.T) {
-	tests := []struct {
-		name string
-		k    FilterKind
-		want string
-	}{
-		{name: "basic", k: FilterKind_QNames, want: "QNames"},
-		{name: "out of range", k: FilterKind_count + 1, want: (FilterKind_count + 1).String()},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := tt.k.TrimString(); got != tt.want {
-				t.Errorf("%v.(FilterKind).TrimString() = %v, want %v", tt.k, got, tt.want)
-			}
-		})
-	}
+// Renders an FilterKind in human-readable form, without "FilterKind_" prefix,
+// suitable for debugging or error messages
+func (k FilterKind) TrimString() string {
+	const pref = "FilterKind_"
+	return strings.TrimPrefix(k.String(), pref)
 }
